Change generation type to `NonZeroU64`
From the [docs](https://doc.rust-lang.org/nightly/std/num/struct.NonZeroU64.html): [A `NonZeroU64` is] an integer that is known not to equal zero. This enables some memory layout optimization. For example, `Option<NonZeroU64>` is the same size as `u64`.

[This example](https://play.rust-lang.org/?version=stable&mode=debug&edition=2018&gist=764d7db69f547b3bef75164bfb978660) shows that `Option<Index>` takes 24 bytes, but by this optimization it would take 16 bytes like a plain `Index`.

I run the benchmarks and the results are __almost__ the same. A small regression (~3% on one test) is probably caused by the fact that incrementing a `NonZeroU64` by 1 requires converting to a `u64` and back. If the type gets better support by the standard library, that may be optimized further.

Still, memory optimization is significative. I myself may need to store a large amount of `Option<Index>` and I belive it's a rather common use-case.